requirement.txt
langchain
langchain-community
boto3
python-dotenv
langchain-aws
langchain-anthropic
langchain-openai